README: Rename 'samples' to 'examples'
At first glance, “example” and “sample” appear convey the same meaning, but there are slight yet important differences between the two concepts.

Examples are used to exemplify and illustrate something. “Example” is also utilized as a tool for the explanation and reinforcement of a particular point. Moreover, examples are used for strict compliance or as a premeditated experience. In this manner, it is expected that the example will be followed and replicated among its audience.

On the other hand, a sample is a small part of something much bigger. Unlike an example, a sample is random and not specific. Samples are often used to describe the quality or nature of a specific whole. “Sample” is often used in statistics or quantitative research as a term to describe part of a target population. Samples are often tangible parts and can be observed using the five senses of sight, touch, taste, smell, and hearing.

See http://www.differencebetween.net/language/words-language/difference-between-example-and-sample/

Signed-off-by: Thomas Steenbergen <thomas.steenbergen@here.com>